Update User entity && repository

+++ backend/src/main/java/be/unamur/info/mdl/groggy/dal/UserRepository.java
@@ -3,11 +3,11 @@ package be.unamur.info.mdl.groggy.dal;
 
 import org.springframework.data.repository.CrudRepository;
 
-import entity.User;
+import be.unamur.info.mdl.groggy.dal.entity.User;
 
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 
 }
